+++ extensions/tmtv/src/Panels/PanelROIThresholdSegmentation/index.ts
@@ -1,0 +1,3 @@
+import PanelROIThresholdSegmentation from './PanelROIThresholdSegmentation';
+
+export default PanelROIThresholdSegmentation;
