+++ modes/tmtv/src/toolbarButtons.js
@@ -250,8 +250,10 @@ const toolbarButtons = [
         {
           commandName: 'displayNotification',
           commandOptions: {
+            title: 'RectangleROI Threshold Tip',
             text:
               'RectangleROI Threshold tool should be used on PT Axial Viewport',
+            type: 'info',
           },
         },
         {
